requirements.txt
Flask
Flask-Smorest
SQLAlchemy
psycopg2-binary
marshmallow
pytest
alembic
